fix: More defensive code when trying to drop tables

+++ migration/2_turba_upgrade_autoincrement.php
@@ -21,9 +21,8 @@ class TurbaUpgradeAutoIncrement extends Horde_Db_Migration_Base
     public function up()
     {
         $this->changeColumn('turba_shares', 'share_id', 'autoincrementKey');
-        try {
+        if (in_array('turba_shares_seq', $this->tables())) {
             $this->dropTable('turba_shares_seq');
-        } catch (Horde_Db_Exception $e) {
         }
     }
